requirements.txt
discord.py>=2.0.0
python-dotenv
requests>=2.32.4
bs4
